+++ dogm/demo/utils/include/types.h
@@ -1,0 +1,16 @@
+// Copyright (c) 2020 Michael Koesel and respective contributors
+// SPDX-License-Identifier: MIT
+// See accompanying LICENSE file for detailed information
+
+#ifndef TYPES_H
+#define TYPES_H
+
+struct PointWithVelocity
+{
+    float x{0.0f};
+    float y{0.0f};
+    float v_x{0.0f};
+    float v_y{0.0f};
+};
+
+#endif  // TYPES_H
